Load Constructor Info

Loads constructor info for all participants in a given season. Use .load_constructors()for an uncached version of this function

load_constructors()

Value

A tibble with one row per constructor

Examples

#Load the list of all constructors who have participated in a F1 race:
load_constructors()
#>            constructorId                      name   nationality
#> 1                  adams                     Adams      American
#> 2                    afm                       AFM        German
#> 3                    ags                       AGS        French
#> 4                   alfa                Alfa Romeo         Swiss
#> 5             alphatauri                AlphaTauri       Italian
#> 6                 alpine            Alpine F1 Team        French
#> 7                   alta                      Alta       British
#> 8                   amon                      Amon New Zealander
#> 9                   moda               Andrea Moda       Italian
#> 10               apollon                   Apollon         Swiss
#> 11                arrows                    Arrows       British
#> 12        arzani-volpini            Arzani-Volpini       Italian
#> 13           butterworth         Aston Butterworth       British
#> 14          aston_martin              Aston Martin       British
#> 15                   ats                       ATS       Italian
#> 16                   bar                       BAR       British
#> 17         behra-porsche             Behra-Porsche       Italian
#> 18               bellasi                   Bellasi         Swiss
#> 19              benetton                  Benetton       Italian
#> 20                   bmw                       BMW        German
#> 21            bmw_sauber                BMW Sauber        German
#> 22                  boro                      Boro         Dutch
#> 23               brabham                   Brabham       British
#> 24    brabham-alfa_romeo        Brabham-Alfa Romeo       British
#> 25           brabham-brm               Brabham-BRM       British
#> 26        brabham-climax            Brabham-Climax       British
#> 27          brabham-ford              Brabham-Ford       British
#> 28         brabham-repco             Brabham-Repco       British
#> 29                 brawn                     Brawn       British
#> 30                   brm                       BRM       British
#> 31              brm-ford                  BRM-Ford       British
#> 32                bromme                    Bromme      American
#> 33                   brp                       BRP       British
#> 34               bugatti                   Bugatti        French
#> 35              caterham                  Caterham     Malaysian
#> 36           vhristensen               Christensen      American
#> 37             cisitalia                 Cisitalia       Italian
#> 38                coloni                    Coloni       Italian
#> 39             connaught                 Connaught       British
#> 40                connew                    Connew       British
#> 41                cooper                    Cooper       British
#> 42     cooper-alfa_romeo         Cooper-Alfa Romeo       British
#> 43            cooper-ats                Cooper-ATS       British
#> 44       cooper-borgward           Cooper-Borgward       British
#> 45            cooper-brm                Cooper-BRM       British
#> 46    cooper-castellotti        Cooper-Castellotti       British
#> 47         cooper-climax             Cooper-Climax       British
#> 48        cooper-ferrari            Cooper-Ferrari       British
#> 49           cooper-ford               Cooper-Ford       British
#> 50       cooper-maserati           Cooper-Maserati       British
#> 51           cooper-osca               Cooper-OSCA       British
#> 52               dallara                   Dallara       Italian
#> 53                tomaso                 De Tomaso       Italian
#> 54  de_tomaso-alfa_romeo      De Tomaso-Alfa Romeo       Italian
#> 55     de_tomaso-ferrari         De Tomaso-Ferrari       Italian
#> 56        de_tomaso-osca            De Tomaso-Osca       Italian
#> 57                 deidt                     Deidt      American
#> 58               del_roy                   Del Roy      American
#> 59            derrington                Derrington       British
#> 60                  dunn                      Dunn      American
#> 61                 eagle                     Eagle      American
#> 62          eagle-climax              Eagle-Climax      American
#> 63         eagle-weslake             Eagle-Weslake      American
#> 64                 elder                     Elder      American
#> 65                  hill              Embassy Hill       British
#> 66              emeryson                  Emeryson       British
#> 67                   emw                       EMW   East German
#> 68                   enb                       ENB       Belgian
#> 69                ensign                    Ensign       British
#> 70               epperly                   Epperly      American
#> 71                   era                       ERA       British
#> 72              eurobrun                 Euro Brun       Italian
#> 73                 ewing                     Ewing      American
#> 74              ferguson                  Ferguson       British
#> 75               ferrari                   Ferrari       Italian
#> 76            fittipaldi                Fittipaldi     Brazilian
#> 77             fondmetal                 Fondmetal       Italian
#> 78              footwork                  Footwork       British
#> 79           force_india               Force India        Indian
#> 80                 forti                     Forti       Italian
#> 81           frazer_nash               Frazer Nash       British
#> 82                   fry                       Fry       British
#> 83                 gilby                     Gilby       British
#> 84               gordini                   Gordini        French
#> 85                  haas              Haas F1 Team      American
#> 86                  hall                      Hall      American
#> 87               hesketh                   Hesketh       British
#> 88                 honda                     Honda      Japanese
#> 89                   hrt                       HRT       Spanish
#> 90                   hwm                       HWM       British
#> 91          iso_marlboro              Iso Marlboro       British
#> 92                jaguar                    Jaguar       British
#> 93                   jbw                       JBW       British
#> 94                jordan                    Jordan         Irish
#> 95               kauhsen                   Kauhsen        German
#> 96                 klenk                     Klenk        German
#> 97                kojima                    Kojima      Japanese
#> 98          kurtis_kraft              Kurtis Kraft      American
#> 99                 kuzma                     Kuzma      American
#> 100                lambo                     Lambo       Italian
#> 101               lancia                    Lancia       Italian
#> 102              langley                   Langley      American
#> 103            larrousse                 Larrousse        French
#> 104                  lds                       LDS South African
#> 105       lds-alfa_romeo            LDS-Alfa Romeo South African
#> 106           lds-climax                LDS-Climax South African
#> 107                  lec                       LEC       British
#> 108             lesovsky                  Lesovsky      American
#> 109               leyton              Leyton House       British
#> 110                 life                      Life       Italian
#> 111               ligier                    Ligier        French
#> 112                 lola                      Lola       British
#> 113         lotus_racing                     Lotus     Malaysian
#> 114             lotus_f1                  Lotus F1       British
#> 115       lotus-borgward            Lotus-Borgward       British
#> 116            lotus-brm                 Lotus-BRM       British
#> 117         lotus-climax              Lotus-Climax       British
#> 118           lotus-ford                Lotus-Ford       British
#> 119       lotus-maserati            Lotus-Maserati       British
#> 120             lotus-pw Lotus-Pratt &amp; Whitney       British
#> 121               lyncar                    Lyncar       British
#> 122                 maki                      Maki      Japanese
#> 123                manor            Manor Marussia       British
#> 124                march                     March       British
#> 125     march-alfa_romeo          March-Alfa Romeo       British
#> 126           march-ford                March-Ford       British
#> 127             marchese                  Marchese      American
#> 128              martini                   Martini        French
#> 129             marussia                  Marussia       Russian
#> 130             maserati                  Maserati       Italian
#> 131                matra                     Matra        French
#> 132           matra-ford                Matra-Ford        French
#> 133                  mbm                       MBM         Swiss
#> 134              mcguire                   McGuire    Australian
#> 135              mclaren                   McLaren       British
#> 136   mclaren-alfa_romeo        McLaren-Alfa Romeo       British
#> 137          mclaren-brm               McLaren-BRM       British
#> 138         mclaren-ford              McLaren-Ford       British
#> 139        mclaren-seren       McLaren-Serenissima       British
#> 140             mercedes                  Mercedes        German
#> 141             merzario                  Merzario       Italian
#> 142            meskowski                 Meskowski      American
#> 143                  mf1                       MF1       Russian
#> 144               milano                    Milano       Italian
#> 145              minardi                   Minardi       Italian
#> 146                moore                     Moore      American
#> 147              nichels                   Nichels      American
#> 148                olson                     Olson      American
#> 149                 onyx                      Onyx       British
#> 150                 osca                      OSCA       Italian
#> 151               osella                    Osella       Italian
#> 152              pacific                   Pacific       British
#> 153             pankratz                  Pankratz      American
#> 154             parnelli                  Parnelli      American
#> 155                 pawl                      Pawl      American
#> 156               penske                    Penske      American
#> 157             phillips                  Phillips      American
#> 158             politoys                  Politoys       British
#> 159              porsche                   Porsche        German
#> 160                prost                     Prost        French
#> 161               protos                    Protos       British
#> 162         racing_point              Racing Point       British
#> 163                  rae                       Rae      American
#> 164                  ram                       RAM       British
#> 165                   re                        RE     Rhodesian
#> 166              rebaque                   Rebaque       Mexican
#> 167             red_bull                  Red Bull      Austrian
#> 168              renault                   Renault        French
#> 169                 rial                      Rial        German
#> 170               sauber                    Sauber         Swiss
#> 171               scarab                    Scarab      American
#> 172            schroeder                 Schroeder      American
#> 173             scirocco                  Scirocco       British
#> 174               shadow                    Shadow       British
#> 175          shadow-ford               Shadow-Ford       British
#> 176         shadow-matra              Shadow-Matra       British
#> 177              shannon                   Shannon       British
#> 178              sherman                   Sherman      American
#> 179                simca                     Simca        French
#> 180               simtek                    Simtek       British
#> 181           snowberger                Snowberger      American
#> 182               spirit                    Spirit       British
#> 183               spyker                    Spyker         Dutch
#> 184           spyker_mf1                Spyker MF1         Dutch
#> 185               stebro                    Stebro      Canadian
#> 186              stevens                   Stevens      American
#> 187              stewart                   Stewart       British
#> 188          super_aguri               Super Aguri      Japanese
#> 189              surtees                   Surtees       British
#> 190               sutton                    Sutton      American
#> 191                 lago               Talbot-Lago        French
#> 192           team_lotus                Team Lotus       British
#> 193              tec-mec                   Tec-Mec       Italian
#> 194                tecno                     Tecno       Italian
#> 195             theodore                  Theodore     Hong Kong
#> 196                token                     Token       British
#> 197              toleman                   Toleman       British
#> 198           toro_rosso                Toro Rosso       Italian
#> 199               toyota                    Toyota      Japanese
#> 200               trevis                    Trevis      American
#> 201               trojan                    Trojan       British
#> 202               turner                    Turner      American
#> 203              tyrrell                   Tyrrell       British
#> 204              vanwall                   Vanwall       British
#> 205              veritas                   Veritas        German
#> 206               virgin                    Virgin       British
#> 207               watson                    Watson      American
#> 208            wetteroth                 Wetteroth      American
#> 209             williams                  Williams       British
#> 210                 wolf                      Wolf      Canadian
#> 211             zakspeed                  Zakspeed        German