Load Driver Info

Loads driver info for all participants in a given season. Use .load_drivers() for an uncached version of this function.

load_drivers(season = get_current_season())

Arguments

season

number from 1950 to current season (defaults to current season).

Value

A tibble with columns driver_id (unique and recurring), first name, last name, nationality, date of birth (yyyy-mm-dd format), driver code, and permanent number (for post-2014 drivers).

Examples

# Load the drivers appearing in this season's data:
load_drivers()
#> # A tibble: 20 × 7
#>    driver_id       given_name family_name nationality date_of_birth code 
#>    <chr>           <chr>      <chr>       <chr>       <chr>         <chr>
#>  1 albon           Alexander  Albon       Thai        1996-03-23    ALB  
#>  2 alonso          Fernando   Alonso      Spanish     1981-07-29    ALO  
#>  3 bottas          Valtteri   Bottas      Finnish     1989-08-28    BOT  
#>  4 de_vries        Nyck       de Vries    Dutch       1995-02-06    DEV  
#>  5 gasly           Pierre     Gasly       French      1996-02-07    GAS  
#>  6 hamilton        Lewis      Hamilton    British     1985-01-07    HAM  
#>  7 hulkenberg      Nico       Hülkenberg  German      1987-08-19    HUL  
#>  8 leclerc         Charles    Leclerc     Monegasque  1997-10-16    LEC  
#>  9 kevin_magnussen Kevin      Magnussen   Danish      1992-10-05    MAG  
#> 10 norris          Lando      Norris      British     1999-11-13    NOR  
#> 11 ocon            Esteban    Ocon        French      1996-09-17    OCO  
#> 12 perez           Sergio     Pérez       Mexican     1990-01-26    PER  
#> 13 piastri         Oscar      Piastri     Australian  2001-04-06    PIA  
#> 14 russell         George     Russell     British     1998-02-15    RUS  
#> 15 sainz           Carlos     Sainz       Spanish     1994-09-01    SAI  
#> 16 sargeant        Logan      Sargeant    American    2000-12-31    SAR  
#> 17 stroll          Lance      Stroll      Canadian    1998-10-29    STR  
#> 18 tsunoda         Yuki       Tsunoda     Japanese    2000-05-11    TSU  
#> 19 max_verstappen  Max        Verstappen  Dutch       1997-09-30    VER  
#> 20 zhou            Guanyu     Zhou        Chinese     1999-05-30    ZHO  
#> # ℹ 1 more variable: permanent_number <chr>

# Load the drivers who participated in the first year of Formula 1:
load_drivers(1950)
#> # A tibble: 40 × 5
#>    driver_id    given_name family_name  nationality date_of_birth
#>    <chr>        <chr>      <chr>        <chr>       <chr>        
#>  1 ader         Walt       Ader         American    1913-12-15   
#>  2 agabashian   Fred       Agabashian   American    1913-08-21   
#>  3 ascari       Alberto    Ascari       Italian     1918-07-13   
#>  4 banks        Henry      Banks        American    1913-06-14   
#>  5 bettenhausen Tony       Bettenhausen American    1916-09-12   
#>  6 biondetti    Clemente   Biondetti    Italian     1898-08-18   
#>  7 bira         Prince     Bira         Thai        1914-07-15   
#>  8 bonetto      Felice     Bonetto      Italian     1903-06-09   
#>  9 branca       Toni       Branca       Swiss       1916-09-15   
#> 10 walt_brown   Walt       Brown        American    1911-12-30   
#> # ℹ 30 more rows