Load Sprint Results

Loads final race results for a given year and round. Note not all rounds have sprint results. Use .load_sprint() for an uncached version of this function.

load_sprint(season = get_current_season(), round = "last")

Arguments

season

number from 2021 to current season (defaults to current season).

round

number from 1 to 23 (depending on season), and defaults to most recent. Also accepts 'last'.

Value

A dataframetibble with columns driver_id, constructor_id, points awarded, finishing position, grid position, laps completed, race status (finished or otherwise), gap to first place, fastest lap, fastest lap time, fastest lap in seconds, or NULL if no sprint exists for this season/round combo

Examples

# Load a sprint result
load_sprint(2021, 10)
#> # A tibble: 20 × 11
#>    driver_id       constructor_id points position grid  laps  status gap   lap  
#>    <chr>           <chr>          <chr>  <chr>    <chr> <chr> <chr>  <chr> <chr>
#>  1 max_verstappen  red_bull       3      1        2     17    Finis… 25:3… 14   
#>  2 hamilton        mercedes       2      2        1     17    Finis… +1.4… 17   
#>  3 bottas          mercedes       1      3        3     17    Finis… +7.5… 17   
#>  4 leclerc         ferrari        0      4        4     17    Finis… +11.… 16   
#>  5 norris          mclaren        0      5        6     17    Finis… +24.… 16   
#>  6 ricciardo       mclaren        0      6        7     17    Finis… +30.… 17   
#>  7 alonso          alpine         0      7        11    17    Finis… +43.… 17   
#>  8 vettel          aston_martin   0      8        10    17    Finis… +44.… 17   
#>  9 russell         williams       0      9        8     17    Finis… +46.… 17   
#> 10 ocon            alpine         0      10       13    17    Finis… +47.… 16   
#> 11 sainz           ferrari        0      11       9     17    Finis… +47.… 10   
#> 12 gasly           alphatauri     0      12       12    17    Finis… +48.… 17   
#> 13 raikkonen       alfa           0      13       17    17    Finis… +50.… 12   
#> 14 stroll          aston_martin   0      14       15    17    Finis… +52.… 12   
#> 15 giovinazzi      alfa           0      15       14    17    Finis… +53.… 14   
#> 16 tsunoda         alphatauri     0      16       16    17    Finis… +53.… 14   
#> 17 latifi          williams       0      17       18    17    Finis… +55.… 17   
#> 18 mick_schumacher haas           0      18       19    17    Finis… +1:0… 15   
#> 19 mazepin         haas           0      19       20    17    Finis… +1:1… 3    
#> 20 perez           red_bull       0      20       5     16    Vibra… NA    6    
#> # ℹ 2 more variables: fastest <chr>, time_sec <dbl>

# If a weekend doesn't have sprint race/quali results, it will return NULL
load_sprint(2022, 1)
#> ! No Sprint data for season = 2022, round = 1
#> NULL