Setup fastf1 connection

Installs or optionally updates fastf1 Python package in the current active Python environment/virtualenv/conda env.

More information on how to manage complex environment needs can be read in the reticulate docs, and tools for managing virtual environments are documented in virtualenv-tools and conda-tools

Usage

setup_fastf1(
  ...,
  envname = "f1dataR_env",
  new_env = identical(envname, "f1dataR_env")
)

Arguments

...

Additional parameters to pass to py_install

envname

Optionally pass an environment name used. Defaults to package default of f1dataR_env.

new_env

Whether or not to completely remove and recreate the environment provided in envname. This will fix any issues experienced by fastf1 related to package dependencies.

Value

No return value, called to install or update fastf1 Python package.

Examples

if (FALSE) { # \dontrun{
# Install fastf1 into the currently active Python environment
setup_fastf1()

# Reinstall fastf1 and recreate the environment.
setup_fastf1(envname = "f1dataR_env", new_env = TRUE)
} # }